+++ transaction-status/src/lib.rs
@@ -854,158 +854,7 @@ pub struct TransactionByAddrInfo {
 
 #[cfg(test)]
 mod test {
-    use {super::*, serde_json::json};
-
-    #[test]
-    fn test_decode_invalid_transaction() {
-        // This transaction will not pass sanitization
-        let unsanitary_transaction = EncodedTransaction::Binary(
-            "ju9xZWuDBX4pRxX2oZkTjxU5jB4SSTgEGhX8bQ8PURNzyzqKMPPpNvWihx8zUe\
-             FfrbVNoAaEsNKZvGzAnTDy5bhNT9kt6KFCTBixpvrLCzg4M5UdFUQYrn1gdgjX\
-             pLHxcaShD81xBNaFDgnA2nkkdHnKtZt4hVSfKAmw3VRZbjrZ7L2fKZBx21CwsG\
-             hD6onjM2M3qZW5C8J6d1pj41MxKmZgPBSha3MyKkNLkAGFASK"
-                .to_string(),
-            TransactionBinaryEncoding::Base58,
-        );
-        assert!(unsanitary_transaction.decode().is_none());
-    }
-
-    #[test]
-    fn test_satisfies_commitment() {
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: None,
-            status: Ok(()),
-            err: None,
-            confirmation_status: Some(TransactionConfirmationStatus::Finalized),
-        };
-
-        assert!(status.satisfies_commitment(CommitmentConfig::finalized()));
-        assert!(status.satisfies_commitment(CommitmentConfig::confirmed()));
-        assert!(status.satisfies_commitment(CommitmentConfig::processed()));
-
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: Some(10),
-            status: Ok(()),
-            err: None,
-            confirmation_status: Some(TransactionConfirmationStatus::Confirmed),
-        };
-
-        assert!(!status.satisfies_commitment(CommitmentConfig::finalized()));
-        assert!(status.satisfies_commitment(CommitmentConfig::confirmed()));
-        assert!(status.satisfies_commitment(CommitmentConfig::processed()));
-
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: Some(1),
-            status: Ok(()),
-            err: None,
-            confirmation_status: Some(TransactionConfirmationStatus::Processed),
-        };
-
-        assert!(!status.satisfies_commitment(CommitmentConfig::finalized()));
-        assert!(!status.satisfies_commitment(CommitmentConfig::confirmed()));
-        assert!(status.satisfies_commitment(CommitmentConfig::processed()));
-
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: Some(0),
-            status: Ok(()),
-            err: None,
-            confirmation_status: None,
-        };
-
-        assert!(!status.satisfies_commitment(CommitmentConfig::finalized()));
-        assert!(!status.satisfies_commitment(CommitmentConfig::confirmed()));
-        assert!(status.satisfies_commitment(CommitmentConfig::processed()));
-
-        // Test single_gossip fallback cases
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: Some(1),
-            status: Ok(()),
-            err: None,
-            confirmation_status: None,
-        };
-        assert!(!status.satisfies_commitment(CommitmentConfig::confirmed()));
-
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: Some(2),
-            status: Ok(()),
-            err: None,
-            confirmation_status: None,
-        };
-        assert!(status.satisfies_commitment(CommitmentConfig::confirmed()));
-
-        let status = TransactionStatus {
-            slot: 0,
-            confirmations: None,
-            status: Ok(()),
-            err: None,
-            confirmation_status: None,
-        };
-        assert!(status.satisfies_commitment(CommitmentConfig::confirmed()));
-    }
-
-    #[test]
-    fn test_serde_empty_fields() {
-        fn test_serde<'de, T: serde::Serialize + serde::Deserialize<'de>>(
-            json_input: &'de str,
-            expected_json_output: &str,
-        ) {
-            let typed_meta: T = serde_json::from_str(json_input).unwrap();
-            let reserialized_value = json!(typed_meta);
-
-            let expected_json_output_value: serde_json::Value =
-                serde_json::from_str(expected_json_output).unwrap();
-            assert_eq!(reserialized_value, expected_json_output_value);
-        }
-
-        let json_input = "{\
-            \"err\":null,\
-            \"status\":{\"Ok\":null},\
-            \"fee\":1234,\
-            \"preBalances\":[1,2,3],\
-            \"postBalances\":[4,5,6]\
-        }";
-        let expected_json_output = "{\
-            \"err\":null,\
-            \"status\":{\"Ok\":null},\
-            \"fee\":1234,\
-            \"preBalances\":[1,2,3],\
-            \"postBalances\":[4,5,6],\
-            \"innerInstructions\":null,\
-            \"logMessages\":null,\
-            \"preTokenBalances\":null,\
-            \"postTokenBalances\":null,\
-            \"rewards\":null\
-        }";
-        test_serde::<UiTransactionStatusMeta>(json_input, expected_json_output);
-
-        let json_input = "{\
-            \"accountIndex\":5,\
-            \"mint\":\"DXM2yVSouSg1twmQgHLKoSReqXhtUroehWxrTgPmmfWi\",\
-            \"uiTokenAmount\": {
-                \"amount\": \"1\",\
-                \"decimals\": 0,\
-                \"uiAmount\": 1.0,\
-                \"uiAmountString\": \"1\"\
-            }\
-        }";
-        let expected_json_output = "{\
-            \"accountIndex\":5,\
-            \"mint\":\"DXM2yVSouSg1twmQgHLKoSReqXhtUroehWxrTgPmmfWi\",\
-            \"uiTokenAmount\": {
-                \"amount\": \"1\",\
-                \"decimals\": 0,\
-                \"uiAmount\": 1.0,\
-                \"uiAmountString\": \"1\"\
-            }\
-        }";
-        test_serde::<UiTransactionTokenBalance>(json_input, expected_json_output);
-    }
+    use super::*;
 
     #[test]
     fn test_ui_transaction_status_meta_ctors_serialization() {
